refactor: reorganize TUI package into modular subpackages and update imports
- move chat, code, git, settings, and terminal components into dedicated subpackages
- update all import paths to reflect new structure
- remove unused imports and minor cleanups
