+++ src/jrdev/ui/tui/chat/chat_list.py
