+++ src/jrdev/ui/tui/settings/model_management/base_model_modal.py
