+++ src/jrdev/ui/tui/settings/model_profile_widget.py
