+++ src/jrdev/ui/tui/settings/terminal_styles_widget.py
