+++ src/jrdev/ui/tui/terminal/bordered_switcher.py
