+++ src/jrdev/ui/tui/terminal/command_confirmation_widget.py
