+++ src/jrdev/ui/tui/terminal/terminal_text_area.py
